Move review plugin integration in its own package
