+++ src/test/java/sonia/scm/issuetracker/internal/PersonMapperTest.java
@@ -29,6 +29,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import sonia.scm.issuetracker.internal.PersonMapper;
 import sonia.scm.repository.Person;
 import sonia.scm.user.DisplayUser;
 import sonia.scm.user.UserDisplayManager;
