+++ src/test/java/sonia/scm/issuetracker/internal/review/PullRequestMapperTest.java
@@ -22,7 +22,7 @@
  * SOFTWARE.
  */
 
-package sonia.scm.issuetracker.internal;
+package sonia.scm.issuetracker.internal.review;
 
 import com.cloudogu.scm.review.pullrequest.service.PullRequest;
 import org.junit.jupiter.api.BeforeEach;
@@ -32,17 +32,14 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import sonia.scm.config.ScmConfiguration;
 import sonia.scm.issuetracker.api.IssueReferencingObject;
+import sonia.scm.issuetracker.internal.PersonMapper;
 import sonia.scm.repository.Repository;
 import sonia.scm.repository.RepositoryTestData;
-import sonia.scm.user.DisplayUser;
 import sonia.scm.user.UserDisplayManager;
-import sonia.scm.user.UserTestData;
 
 import java.time.Instant;
-import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.Mockito.when;
 
 
 @ExtendWith(MockitoExtension.class)
